+++ kepler-core/src/models/kv_delete.rs
@@ -10,12 +10,12 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub orbit: String,
     #[sea_orm(primary_key)]
-    pub seq: u32,
+    pub seq: i64,
     #[sea_orm(primary_key)]
     pub epoch_id: Hash,
 
     pub key: String,
-    pub deleted_seq: u32,
+    pub deleted_seq: i64,
     pub deleted_epoch_id: Hash,
 }
